requirements.txt
pytest-django
requests
pillow
sorl-thumbnail
django
djangorestframework
isort==4.3.21

